refactor(persistence): extract order data models from orders_store
orders_store.py (870 lines) mixed the OrderStatus lifecycle enum and OrderRecord
data contract (with its (de)serialization and checksum helpers) with the
SQLite-backed OrdersStore machinery. Split the contracts into
persistence/orders_models.py, leaving orders_store to own the connection
handling, schema, and store behavior.

orders_store re-exports OrderRecord/OrderStatus, so all 31 import sites are
byte-stable with identical class identity (verified, incl. to_dict/from_dict
round-trip). No behavior change.
